camfusion/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
httpx==0.28.1
Pillow==11.1.0
numpy==2.2.4
